Entity-based events WIP

+++ src/runtime/api/plugins/entity/components/vector_2d.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub(crate) struct Vector2D {
-    x: u32,
-    y: u32,
+    x: f32,
+    y: f32,
 }
